backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
email-validator==2.2.0
pydantic-settings==2.4.0
sqlalchemy==2.0.32
psycopg[binary]==3.2.13
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
gTTS==2.5.3
email-validator==2.2.0
